implented unit tests
